requirements.txt
fastapi
uvicorn[standard]
sqlmodel
google-genai
python-dotenv
requests
beautifulsoup4
streamlit
